Reorg readTable functions
