+++ javascript/tables/src/index.ts
@@ -14,7 +14,7 @@ export * from './guessDelimiter.js'
 export * from './introspect.js'
 export * from './parseCSV.js'
 export * from './parseTypes.js'
-export { readTable } from './readTable.js'
+export * from './readTable.js'
 export * from './rows.js'
 export * from './stats.js'
 export * from './table.js'
