+++ javascript/tables/src/readCsvTable.constants.ts
@@ -3,8 +3,7 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 
-import type { DataShape, ParserOptions } from '@datashaper/schema'
-import { DataOrientation } from '@datashaper/schema'
+import type { ParserOptions } from '@datashaper/schema'
 
 import { COMMENT_DEFAULT, DELIMITER_DEFAULT } from './typeHints.defaults.js'
 
@@ -29,7 +28,6 @@ export const PAPAPARSE_SUPPORTED_OPTS = new Set([
 	'escapeChar',
 	'skipBlankLines',
 	'skipRows',
-	'readRows',
 ])
 
 export const PAPAPARSE_PROPS_MAP = {
@@ -44,8 +42,4 @@ export const LINE_TERMINATORS = new Set(['\r', '\r\n', '\n'])
 export const DEFAULT_PARSER_OPTIONS: ParserOptions = {
 	delimiter: DELIMITER_DEFAULT,
 	comment: COMMENT_DEFAULT,
-}
-
-export const DEFAULT_DATA_SHAPE_JSON_OPTIONS: DataShape = {
-	orientation: DataOrientation.Records,
 }
