+++ javascript/tables/src/readCsvTable.types.ts
